Added apt and dnf into the mix
This commit provides support for both package managers but there are rough
edges everywhere, so more commits will come to fix this.

Moreover, it's still not clear how we should handle the CLI (e.g. the "patch"
concept does not make a lot of sense in neither dnf nor apt)

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>